review: adding route test to the right place
